Implement store; Move reset and restart into application
