+++ modules/windows/AppWindow.js
@@ -1,5 +1,5 @@
 const { ipcMain, BrowserWindow, app } = require('electron');
-const store = require('./store').get();
+const store = require('../store').get();
 
 let { WindowList } = require('./index');
