+++ modules/windows/SettingsWindow.js
@@ -1,5 +1,5 @@
 const { ipcMain, BrowserWindow } = require('electron');
-const store = require('./store').get();
+const store = require('../store').get();
 
 const { DEFAULT_INTERVAL } = require('../constants');
 let { WindowList } = require('./index');
